Move the MenuType into a separate file.
